changed { convention. updated Force class to account for velocity. ready to add collision detection tommorow.

+++ src/Machine.hpp
@@ -4,15 +4,16 @@
 #include "Component.hpp"
 #include <vector>
 
-class Machine {
+class Machine
+{
 	double perspective_point[3];
 	std::vector<Component::Instance> components;
-	std::vector<float> colour_scheme;
+	std::vector<int> colour_scheme;
 public:
 	void draw();
 	void iterate();
 	void pan(double [3]);
 	void add(Component::Instance);
 	void bind();
-	Machine(double [3], std::vector<float>);
+	Machine(double [3], std::vector<int>);
 };
